separates Home component from App
